Enhance accessibiity of warp menu by keyboard
The warp menu is opened by a button on the right-hand side of the screen.
This button was implemented as an empty link-element. To make it more
accessible I changed this element to a button.
This button can be triggered by the keyboard (space and enter). To further enhance accessibility for visually challenged persons I added aria-labels.
I simplified the styling to work with chrome and ie browsers.

resolves #22