Removed static support entries in warpmenu.
Removed static Support entries as they now come with the regular entries as part of the menu.json. The new support entries are configurable in the ces-confd and toggleable via a etcd key.
For example, if you want to remove the support entries for "myClodugou" and "About Cloudogu", you can do so with the command: `etcdctl set /config/_global/disabled_warpmenu_support_entries '["myCloudogu", "aboutCloudoguToken"]'`.

Reworked the translation of the support entries (added testdata in menu.json).
Update Changelog